moved declaration of Chunk into common.
