+++ client/src/graphics/voxels/mod.rs
@@ -326,17 +326,3 @@ impl Loadable for ChunkDesc {
     }
 }
 
-pub enum Chunk {
-    Fresh,
-    Generating,
-    Populated {
-        voxels: VoxelData,
-        surface: Option<SlotId>,
-    },
-}
-
-impl Default for Chunk {
-    fn default() -> Self {
-        Chunk::Fresh
-    }
-}
