+++ common/src/chunk.rs
@@ -1,0 +1,14 @@
+pub enum Chunk {
+    Fresh,
+    Generating,
+    Populated {
+        voxels: VoxelData,
+        surface: Option<SlotId>,
+    },
+}
+
+impl Default for Chunk {
+    fn default() -> Self {
+        Chunk::Fresh
+    }
+}
